Replace ScaleFactorHelper.h with new SF classes
